pyproject.toml
[tool.poetry]
name = "ai-chat"
version = "0.1.0"
description = "FastAPI + Postgres + SQLAlchemy (async) scaffold using Poetry and Docker Compose."
authors = ["chat_api <sandup>"]
readme = "README.md"
packages = [{ include = "chat_api" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
ruff = "^0.8.4"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"


